+++ src/markehme/factionsplus/listeners/PowerboostListener.java
@@ -21,28 +21,17 @@ public class PowerboostListener implements Listener{
 	public void onEntityDeath(EntityDeathEvent event)
 	{
 		
-		//ok this worked but it's integrated below inside canLosePowerInThisArea() because Factions allows 
-		//powerloss in warzone of nopowerloss worlds, and by commenting this out we also do that, and save a few CPU cycles since
-		//the checks are not done for every mob, but for every player (and for every mob only when extraPowerWhenKillMonster>0)
-		//done: allow warzone to powerloss even in nopowerloss worlds, as does Factions code. But also make sure this works in 1.7 too,
-		//not just in 1.6
-//		
 		//1of2
 		if ((event.getEntity() instanceof Player)) {
 			Player p = (Player)event.getEntity();
 			
-			if (!shouldLosePowerWherePlayerIsAt(p)) {
+			if (!canLosePowerWherePlayerIsAt(p)) {
 				// this check is supposed to be for Player only, maybe move the call for this method below
 				//we don't want this check to happen on every entity death on the server
+//				FactionsPlus.info( "no power loss");
 				return;//this will deny extra gain or loss from below
-			}else {
-				//if it should lose power, then, we still don't allow it in nopowerloss worlds
-				//FIXME: this is no good, since it would not allow warzone in nopowerloss worlds to lose power, but it should!
-				if (Utilities.noPowerLossWorld(p.getWorld())) {
-					return;
-				}
 			}
-//			FactionsPlus.info( "here");
+//			FactionsPlus.info( "would lose power");
 			
 			
 			EntityDamageEvent ldc = event.getEntity().getLastDamageCause();//it can be null, see issue 60
@@ -120,9 +109,8 @@ public class PowerboostListener implements Listener{
 				return;
 			}
 			if ( Config._powerboosts.extraPowerWhenKillMonster._ > 0.0D) {
-				if (shouldLosePowerWherePlayerIsAt(k)
-						&& (!Utilities.noPowerLossWorld(k.getWorld()))
-						) {//this IF is here to save some CPU cycles since likely the 0 value is above
+				// done: Block power GAINS in powerloss disabled regions as well: this is gonna be tricky
+				if (canLosePowerWherePlayerIsAt(k)) {//this IF is here to save some CPU cycles since likely the 0 value is above
 					//allow increase only if player can also lose ie. disallow gain in worldsNoPowerLoss worlds 
 					Utilities.addPower( k, Config._powerboosts.extraPowerWhenKillMonster._);
 				}
@@ -131,26 +119,11 @@ public class PowerboostListener implements Listener{
 	}
 
 	
-	public static final boolean shouldLosePowerWherePlayerIsAt(Player player) {
-		if ( !Config._powerboosts.respectFactionsPluginPowerLossRules._ ) {
-			return true;
-		}
-		return canLosePowerWherePlayerIsAt(player);
-	}
-	
 	public static final boolean canLosePowerWherePlayerIsAt( Player player ) {
-		// TODO: Block power GAINS in powerloss disabled regions as well: this is gonna be tricky
 		Faction factionAtFeet = Board.getFactionAt( new FLocation( player.getLocation() ) );
 		// FPlayer fp=FPlayers.i.get(p);
 		// assert null != fp;
 		return canLosePowerInThisFaction(factionAtFeet, player.getWorld());
-	}
-	
-	public static final boolean shouldLosePowerInThisFaction(Faction faction, World worldName) {//this method is not needed;it's only for debug
-		if ( !Config._powerboosts.respectFactionsPluginPowerLossRules._ ) {
-			return true;
-		}
-		return canLosePowerInThisFaction(faction, worldName);
 	}
 	
 	public static final boolean canLosePowerInThisFaction(Faction faction, World worldName) {